@@ -1,18 +1,24 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
+    ext.kotlin_version = '1.3.30'
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
+        classpath 'com.android.tools.build:gradle:3.6.1'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath 'org.jetbrains.dokka:dokka-android-gradle-plugin:0.9.15'
     }
 }
 
 allprojects {
     repositories {
+        google()
         jcenter()
+        maven { url "https://jitpack.io" }
     }
 }
 
@@ -20,14 +26,14 @@ task clean(type: Delete) {
     delete rootProject.buildDir
 }
 ext {
-    compileSdkVersion = 25
-    buildToolsVersion = "25.0.0"
-    designVersion = "com.android.support:design:25.3.1"
-    v4Version = "com.android.support:support-v4:25.3.1"
-    appcompat = "com.android.support:appcompat-v7:25.3.1"
-    recycleview = "com.android.support:recyclerview-v7:25.3.1"
-    targetSdkVersion = 19
-    minSdkVersion = 14
+    superVersion = "1.0.0"
+    compileSdkVersion = 28
+    buildToolsVersion = "28.0.3"
+    designVersion = "com.google.android.material:material:" + superVersion
+    v4Version = "androidx.legacy:legacy-support-v4:" + superVersion
+    v7AppCompat = "androidx.appcompat:appcompat:" + superVersion
+    targetSdkVersion = 24
+    minSdkVersion = 16
     versionCode = 1
-    versionName = "0.1"
+    versionName = "0.0.1"
 }
