@@ -139,6 +139,7 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
     public static final int FFP_PROP_INT64_BIT_RATE                         = 20100;
     public static final int FFP_PROP_INT64_TCP_SPEED                        = 20200;
     public static final int FFP_PROP_INT64_LATEST_SEEK_LOAD_DURATION        = 20300;
+    public static final int FFP_PROP_INT64_IMMEDIATE_RECONNECT              = 20211;
     //----------------------------------------
 
     @AccessedByNative
@@ -949,6 +950,10 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
         native_finalize();
     }
 
+    public void httphookReconnect() {
+        _setPropertyLong(FFP_PROP_INT64_IMMEDIATE_RECONNECT, 1);
+    }
+
     public void setCacheShare(int share) {
         _setPropertyLong(FFP_PROP_INT64_SHARE_CACHE_DATA, (long)share);
     }
@@ -971,82 +976,82 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
             }
 
             switch (msg.what) {
-                case MEDIA_PREPARED:
-                    player.notifyOnPrepared();
-                    return;
+            case MEDIA_PREPARED:
+                player.notifyOnPrepared();
+                return;
 
-                case MEDIA_PLAYBACK_COMPLETE:
-                    player.stayAwake(false);
+            case MEDIA_PLAYBACK_COMPLETE:
+                player.stayAwake(false);
+                player.notifyOnCompletion();
+                return;
+
+            case MEDIA_BUFFERING_UPDATE:
+                long bufferPosition = msg.arg1;
+                if (bufferPosition < 0) {
+                    bufferPosition = 0;
+                }
+
+                long percent = 0;
+                long duration = player.getDuration();
+                if (duration > 0) {
+                    percent = bufferPosition * 100 / duration;
+                }
+                if (percent >= 100) {
+                    percent = 100;
+                }
+
+                // DebugLog.efmt(TAG, "Buffer (%d%%) %d/%d",  percent, bufferPosition, duration);
+                player.notifyOnBufferingUpdate((int)percent);
+                return;
+
+            case MEDIA_SEEK_COMPLETE:
+                player.notifyOnSeekComplete();
+                return;
+
+            case MEDIA_SET_VIDEO_SIZE:
+                player.mVideoWidth = msg.arg1;
+                player.mVideoHeight = msg.arg2;
+                player.notifyOnVideoSizeChanged(player.mVideoWidth, player.mVideoHeight,
+                        player.mVideoSarNum, player.mVideoSarDen);
+                return;
+
+            case MEDIA_ERROR:
+                DebugLog.e(TAG, "Error (" + msg.arg1 + "," + msg.arg2 + ")");
+                if (!player.notifyOnError(msg.arg1, msg.arg2)) {
                     player.notifyOnCompletion();
-                    return;
+                }
+                player.stayAwake(false);
+                return;
 
-                case MEDIA_BUFFERING_UPDATE:
-                    long bufferPosition = msg.arg1;
-                    if (bufferPosition < 0) {
-                        bufferPosition = 0;
-                    }
+            case MEDIA_INFO:
+                switch (msg.arg1) {
+                    case MEDIA_INFO_VIDEO_RENDERING_START:
+                        DebugLog.i(TAG, "Info: MEDIA_INFO_VIDEO_RENDERING_START\n");
+                        break;
+                }
+                player.notifyOnInfo(msg.arg1, msg.arg2);
+                // No real default action so far.
+                return;
+            case MEDIA_TIMED_TEXT:
+                if (msg.obj == null) {
+                    player.notifyOnTimedText(null);
+                } else {
+                    IjkTimedText text = new IjkTimedText(new Rect(0, 0, 1, 1), (String)msg.obj);
+                    player.notifyOnTimedText(text);
+                }
+                return;
+            case MEDIA_NOP: // interface test message - ignore
+                break;
 
-                    long percent = 0;
-                    long duration = player.getDuration();
-                    if (duration > 0) {
-                        percent = bufferPosition * 100 / duration;
-                    }
-                    if (percent >= 100) {
-                        percent = 100;
-                    }
+            case MEDIA_SET_VIDEO_SAR:
+                player.mVideoSarNum = msg.arg1;
+                player.mVideoSarDen = msg.arg2;
+                player.notifyOnVideoSizeChanged(player.mVideoWidth, player.mVideoHeight,
+                        player.mVideoSarNum, player.mVideoSarDen);
+                break;
 
-                    // DebugLog.efmt(TAG, "Buffer (%d%%) %d/%d",  percent, bufferPosition, duration);
-                    player.notifyOnBufferingUpdate((int)percent);
-                    return;
-
-                case MEDIA_SEEK_COMPLETE:
-                    player.notifyOnSeekComplete();
-                    return;
-
-                case MEDIA_SET_VIDEO_SIZE:
-                    player.mVideoWidth = msg.arg1;
-                    player.mVideoHeight = msg.arg2;
-                    player.notifyOnVideoSizeChanged(player.mVideoWidth, player.mVideoHeight,
-                            player.mVideoSarNum, player.mVideoSarDen);
-                    return;
-
-                case MEDIA_ERROR:
-                    DebugLog.e(TAG, "Error (" + msg.arg1 + "," + msg.arg2 + ")");
-                    if (!player.notifyOnError(msg.arg1, msg.arg2)) {
-                        player.notifyOnCompletion();
-                    }
-                    player.stayAwake(false);
-                    return;
-
-                case MEDIA_INFO:
-                    switch (msg.arg1) {
-                        case MEDIA_INFO_VIDEO_RENDERING_START:
-                            DebugLog.i(TAG, "Info: MEDIA_INFO_VIDEO_RENDERING_START\n");
-                            break;
-                    }
-                    player.notifyOnInfo(msg.arg1, msg.arg2);
-                    // No real default action so far.
-                    return;
-                case MEDIA_TIMED_TEXT:
-                    if (msg.obj == null) {
-                        player.notifyOnTimedText(null);
-                    } else {
-                        IjkTimedText text = new IjkTimedText(new Rect(0, 0, 1, 1), (String)msg.obj);
-                        player.notifyOnTimedText(text);
-                    }
-                    return;
-                case MEDIA_NOP: // interface test message - ignore
-                    break;
-
-                case MEDIA_SET_VIDEO_SAR:
-                    player.mVideoSarNum = msg.arg1;
-                    player.mVideoSarDen = msg.arg2;
-                    player.notifyOnVideoSizeChanged(player.mVideoWidth, player.mVideoHeight,
-                            player.mVideoSarNum, player.mVideoSarDen);
-                    break;
-
-                default:
-                    DebugLog.e(TAG, "Unknown message type " + msg.what);
+            default:
+                DebugLog.e(TAG, "Unknown message type " + msg.what);
             }
         }
     }
@@ -1060,7 +1065,7 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
      */
     @CalledByNative
     private static void postEventFromNative(Object weakThiz, int what,
-                                            int arg1, int arg2, Object obj) {
+            int arg1, int arg2, Object obj) {
         if (weakThiz == null)
             return;
 
@@ -1115,7 +1120,7 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
         int EVENT_WILL_HTTP_OPEN = 0x1;                 // ARG_URL
         int EVENT_DID_HTTP_OPEN = 0x2;                  // ARG_URL, ARG_ERROR, ARG_HTTP_CODE
         int EVENT_WILL_HTTP_SEEK = 0x3;                 // ARG_URL, ARG_OFFSET
-        int EVENT_DID_HTTP_SEEK = 0x4;                  // ARG_URL, ARG_OFFSET, ARG_ERROR, ARG_HTTP_CODE
+        int EVENT_DID_HTTP_SEEK = 0x4;                  // ARG_URL, ARG_OFFSET, ARG_ERROR, ARG_HTTP_CODE, ARG_FILE_SIZE
 
         String ARG_URL = "url";
         String ARG_SEGMENT_INDEX = "segment_index";
@@ -1129,6 +1134,7 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
 
         String ARG_OFFSET = "offset";
         String ARG_HTTP_CODE = "http_code";
+        String ARG_FILE_SIZE = "file_size";
 
         /*
          * @return true if invoke is handled
